+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/spring/annotation/SpringValueProcessor.java
@@ -27,6 +27,7 @@ import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.context.annotation.Bean;
+import org.springframework.lang.NonNull;
 
 /**
  * Spring value processor of field or method which has @Value and xml config placeholders.
@@ -68,7 +69,7 @@ public class SpringValueProcessor extends PolarisProcessor implements BeanFactor
 	}
 
 	@Override
-	public Object postProcessBeforeInitialization(Object bean, String beanName)
+	public Object postProcessBeforeInitialization(Object bean, @NonNull String beanName)
 			throws BeansException {
 		if (polarisConfigProperties.isAutoRefresh()) {
 			super.postProcessBeforeInitialization(bean, beanName);